Export 444 for Host based connections
Experiment with having host based proxying exposed on port 444.

This will forward the host header to the upstream proxy which will then use that to look up from the rules for the tenant. This can potentially enable multiple reverse proxies.

However, if we want to support this seamlessly for customers we still need to find a way to make this easily work with their rules.